Breaking change: Redesigned element parsing
Bump rust-version to 1.80.0

Parsers `item` and `seq` are now replaced with `elem`, which
can take as input both individual items and sequences.

Parsers `until_item` and `until_seq` are now replaced with `until` for
the same reason as above.

Introduced `Searchee` to support the above and to allow for user defined
ways to search for elements.

Added default implementations for `Searchee` for `&[T]`, `T`, `str`, and `char`.

Optimized `item_if`

Simplified SliceLike